some verticality configuration

+++ Assets/grid/tile/tileconfig/tileconfig.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class tileconfig:MonoBehaviour
 {
     [Tooltip("Tile height in gameplay units, not actual physical height")]
-    public int tileHeight;
+    public float tileHeight;
     public Color colour;
     public bool obstruction;
 }
